# Refresh tsconfig and eslint setup