--- a/.npmcheckrc
+++ b/.npmcheckrc
@@ -1,7 +1,0 @@
-{
-  "depcheck": {
-    "ignoreMatches": [
-      "@handlebars/parser"
-    ]
-  }
-}