--- a/.npmrc
+++ b/.npmrc
@@ -1,4 +1,3 @@
-use-node-version=20.9.0
 ignore-workspace-root-check=true
 auto-install-peers=true
 strict-peer-dependencies=true