--- a/.prototools
+++ b/.prototools
@@ -1,2 +1,2 @@
 node = "22"
-pnpm = "latest"
+pnpm = "latest-10"