--- a/.prototools
+++ b/.prototools
@@ -1,0 +1,2 @@
+node = "22"
+pnpm = "latest"