--- a/.watchmanconfig
+++ b/.watchmanconfig
@@ -1,4 +1,0 @@
-{
-  "fsevents_latency": 0.02,
-  "ignore_dirs": ["tmp"]
-}