--- a/bin/build-verify.mjs
+++ b/bin/build-verify.mjs
@@ -1,7 +1,8 @@
-import { globby } from 'globby';
 import { readFile } from 'node:fs/promises';
-import { fileURLToPath } from 'node:url';
 import { resolve } from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+import { globby } from 'globby';
 
 const currentDir = fileURLToPath(import.meta.url);
 const FORBIDDEN = [
@@ -16,7 +17,7 @@ const FORBIDDEN = [
   /**
    * These are for local VM debugging and development, and are not meant to make it to real code
    */
-  /[^.]check\(/,
+  /[^.]check\(/u,
   'CheckInterface',
   'CheckOr',
   'CheckFunction',
@@ -45,7 +46,9 @@ for (let filePath of files) {
   let content = file.toString();
 
   for (let searchFor of FORBIDDEN) {
-    if (content.match(searchFor)) {
+    const match = typeof searchFor === 'string' ? content === searchFor : searchFor.test(content);
+
+    if (match) {
       errors.push({ filePath, found: searchFor });
     }
   }