--- a/bin/clean.mjs
+++ b/bin/clean.mjs
@@ -1,3 +1,11 @@
+import { join } from 'node:path';
+
+import repo, { isRoot } from '@glimmer-workspace/repo-metadata';
 import { rimraf } from 'rimraf';
 
-await rimraf('**/{dist,.turbo,node_modules}/', { glob: true });
+for (const pkg of repo.packages) {
+  if (isRoot(pkg)) continue;
+  await rimraf(join(pkg.root, 'dist'));
+}
+
+// await rimraf('**/{dist,.turbo,node_modules}/', { glob: true });