--- a/bin/unlink-all.mts
+++ b/bin/unlink-all.mts
@@ -1,6 +1,7 @@
 import chalk from 'chalk';
 import { execa } from 'execa';
 import { rimraf } from 'rimraf';
+
 import { packages } from './packages.mjs';
 
 for (const pkg of packages('@glimmer')) {