--- a/guides/workspace/demos/auto-import.gif
+++ b/guides/workspace/demos/auto-import.gif
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:48e4f21badf00542cfb61dfff160aa5941f50771eb60c8a170536ed61a4dad96
+size 3058150