This is a significant overhaul of the configuration setup for the repository, driven largely by the work required to update the repository to flat config and ESLint 9.

---

1. Migrate to ESLint 9 + new config format
2. Migrate to a single tsconfig file for the entire repo

In general, this results in more consistency across the repo, including in one-off packages and tests, but it still gives us flexibility to create special rules if necessary.

As part of this process, a new `repo-metadata` package was created that allows other infrastructure packages to get workspace metadata. This package has a static JSON file with the information about each package in the workspace. This was important in the current PR because it allows our eslint config to be sensitive to information in individual packages
without slowing down the boot time of eslint.

Fully migrating to ESLint 9 required:

- Making the whole repo clean with `eslint .`
- Making `eslint .` in individual packages work (and perhaps move the custom configuration to configs in each package
- Moving the guts of the code into `@glimmer-workspace/eslint-plugin`. This should be **much** more straightforward with ESLint 9.
- Porting all of the ESLint rules from the current setup to the new config, or deciding to ignore them.
- Updating our JSON linting

## Increased Benchmark Isolation and Predictability

Rather than using a workspace package for benchmarks and trying to get `vite build` to emulate the expected behavior, we `pack` the dependencies and refer to them directly in the `package.json`.

This is a little bit more involved than it might seem like it should be because the benchmark has to run on the `main` branch, so it has to be very general and can't use any infrastructure in the workspace itself that could affect the builds.

On the bright side, that constraint pushes us towards treating the benchmark environment as a normal consumer of the normal Glimmer packages, which is good.

The only tiny exception to that rule is that the build script builds the `@glimmer-workspace/benchmark-env` package, even though it's not a public package. This means that the code in `@glimmer-workspace/benchmark-env` needs to make sure it can be built.

The current setup allows us to use the `main` branch's `@glimmer-workspace/benchmark-env` package with the experiment's `krausest` package, so it seems reasonable to build the benchmark env and use it as a traditional package rather than trying to get two
workspaces to somehow mesh together (which is what got us into trouble in the first place).

> [!NOTE]
>
> Getting this working right now requires a little bit of hackery when setting up the current `control` (i.e. `main`) branch, because `main` isn't set up to build the `@glimmer-workspace/benchmark-env`.
>
> Once this PR is merged, we should be able to remove those hacks and rely on vanilla package builds for the entire setup going forward.

## Meta Updater and Repo Scripts

This PR introduces a [meta-updater](https://github.com/pnpm/meta-updater) setup that keeps packages aligned with repository conventions based upon whether they're published, need a build, are test packages, etc.

It further standardized per-package scripts via `turbo`, and created top-level `repo:*` scripts to differentiate repo workflows from standard scripts in each package. This PR also begins to document repo workflows (in `guides/workspace`).

## Inline Notes

This PR adds a `@bandaid` highlighting rule to capture situations where code exists as a temporary workaround. The rule and allows you to document the reason for the workaround so it can be identified and cleaned up later.

<kbd> ![image](https://github.com/user-attachments/assets/c771f1f1-e4e3-4b91-bacf-87f217f0ea2f) </kbs>

I also made more use of an existing `@fixme` annotation.

<kbd> ![image](https://github.com/user-attachments/assets/73ee5358-2999-4321-b9e8-05772859c21e) </kbd>

> I think we should improve our workflow around `fixme`s so that they're linked to issues and perhaps can expire based on certain conditions, but we don't need to block this PR on those improvements.
>
> The `@fixme`s added in this PR are almost entirely attached to `eslint-disable` annotations that didn't exist before because of how inconsistently our previous lint setup applied to specific files. They should be cleaned up as soon as possible, but they represent an improvement to the status quo (not a regression).

## Testing Entry Points

While developing this PR, the tests began to fail because Vite was failing to interpret the
`import.meta.glob`s in `index.html` file as entry points. This resulted in multiple copies of the validator code getting included in the tests (via chunks and via the `.ts` module itself).

To make matters worse, the validator code included in the chunk was missing the code that detects duplicates, so the only failure symptom was an error reporting that the `COMPUTE` symbol wasn't present on an instance of `Tag`.

Being explicit about entry points fixed the problem and made the tests pass. I don't know why Vite doesn't interpret `import.meta.glob` properly in this situation, but I think we can live with duplicating the test glob in `vite.config.mjs` for now.

## Other Improvements

- Document vscode workflow recommendations
- Explain recommended extensions for vscode
- Remove many unused dependencies