@NullVoxPopuli 2025-01-19
Current performance boost on my lil'linux laptop:
```
duration phase estimated improvement -383ms [-528ms to -245ms] OR -7.39% [-10.18% to -4.72%]
renderEnd phase estimated improvement -1ms [-2ms to 0ms] OR -5.11% [-11.49% to -0.12%]
render1000Items1End phase no difference [-30ms to 4ms]
clearItems1End phase no difference [-5ms to 3ms]
render1000Items2End phase estimated improvement -17ms [-35ms to -1ms] OR -6.19% [-12.67% to -0.19%]
clearItems2End phase no difference [-2ms to 2ms]
render5000Items1End phase estimated improvement -117ms [-165ms to -79ms] OR -8.74% [-12.4% to -5.92%]
clearManyItems1End phase no difference [-5ms to 9ms]
render5000Items2End phase estimated improvement -114ms [-150ms to -69ms] OR -9.61% [-12.68% to -5.87%]
clearManyItems2End phase no difference [-16ms to 0ms]
render1000Items3End phase no difference [-42ms to 2ms]
append1000Items1End phase no difference [-58ms to 1ms]
append1000Items2End phase estimated improvement -23ms [-47ms to -2ms] OR -7.43% [-15.36% to -0.65%]
updateEvery10thItem1End phase no difference [-11ms to 1ms]
updateEvery10thItem2End phase no difference [-10ms to 0ms]
selectFirstRow1End phase no difference [-11ms to 0ms]
selectSecondRow1End phase no difference [-16ms to 0ms]
removeFirstRow1End phase no difference [-15ms to 1ms]
removeSecondRow1End phase no difference [-17ms to 0ms]
swapRows1End phase no difference [-1ms to 1ms]
swapRows2End phase no difference [-1ms to 0ms]
clearItems4End phase no difference [-3ms to 5ms]
paint phase no difference [0ms to 1ms]
```

this is very nice!

I wonder how much has to do with there just being less code in the output now